Refactors Header component
