+++ packages/lms/src/lib/components/header/Header.svelte
@@ -1,11 +1,23 @@
-<header>
-  <slot />
-</header>
+<script lang="ts">
+	import { AppBar, drawerStore } from '@skeletonlabs/skeleton';
+	import Logo from './Logo.svelte';
+	import Search from './Search.svelte';
+	import Settings from './Settings.svelte';
 
-<style>
-  header {
-    grid-area: header;
-    display: flex;
-    gap: 0.5rem;
-  }
-</style>
+	function drawerOpen(): void {
+		drawerStore.open();
+	}
+</script>
+
+<AppBar gridColumns="grid-cols-3" slotDefault="place-self-center" slotTrail="place-content-end">
+	<svelte:fragment slot="lead">
+		<div class="flex items-center">
+			<button class="lg:hidden btn btn-sm mr-4" on:click={drawerOpen}>
+				<span class="material-symbols-outlined"> menu </span>
+			</button>
+			<Logo />
+		</div>
+	</svelte:fragment>
+	<Search />
+	<svelte:fragment slot="trail"><Settings /></svelte:fragment>
+</AppBar>
